src/main/java/com/twoonezero/webservice/service/PostService.java
package com.twoonezero.webservice.service;


import com.twoonezero.webservice.domain.posts.PostsRepository;
import com.twoonezero.webservice.dto.posts.PostsMainResponseDto;
import com.twoonezero.webservice.dto.posts.PostsSaveRequestDto;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.stream.Collectors;

@AllArgsConstructor
@Service
public class PostService {
    private PostsRepository postsRepository;

    public Long save(PostsSaveRequestDto dto){
        return postsRepository.save(dto.toEntity()).getId();
    }

    @Transactional(readOnly = true)
    public List<PostsMainResponseDto> findAllDesc() {
        return postsRepository.findAllDesc()
                .map(PostsMainResponseDto::new)
                .collect(Collectors.toList());
    }

}
